Deprecate android.net.rtp/android.net.sip.
The VOIP library provided in the Android platform suffers from a number of
issues and has not been maintained in a long time.

It does not support modern audio codecs, it uses broadcasts which rely
on background receivers being woken up (which won't be), and does not work
on IPV6 networks.


Mark android.net.sip code as deprecated.

This code has not been maintained in a long while and has a number of
pretty big flaws including: lack of IPV6 support, reliance on broadcasts
which are supported to wake up background receivers but don't.

Recommendation is for developers to use any number of other open source
SIP VOIP stacks available in the ecosystem.

It is also not currently supported on all Android devices.

Test: Build / observe API surface deprecation.
Bug: 175910634
Change-Id: Id6c8810b187ce06d7035ba58ad8daf849e194c1f